analytics_api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiomysql==0.3.2
    # via -r requirements.in
alembic==1.18.4
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   starlette
    #   watchfiles
cffi==2.0.0
    # via cryptography
click==8.3.2
    # via uvicorn
cryptography==46.0.7
    # via -r requirements.in
fastapi==0.136.0
    # via -r requirements.in
greenlet==3.4.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.12
    # via anyio
mako==1.3.11
    # via alembic
markupsafe==3.0.3
    # via mako
pycparser==3.0
    # via cffi
pydantic==2.13.3
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.46.3
    # via pydantic
pymysql==1.1.2
    # via aiomysql
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.3
    # via uvicorn
sqlalchemy==2.0.49
    # via
    #   -r requirements.in
    #   alembic
starlette==1.0.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn[standard]==0.45.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
